Fix xb360 auth

+++ include/shared_main.h
@@ -14,7 +14,7 @@ void send_report_to_pc(const void *report, uint8_t len);
 bool ready_for_next_packet(void);
 bool usb_configured(void);
 void receive_report_from_controller(uint8_t const *report, uint16_t len);
-void xinput_controller_connected(uint8_t vid, uint8_t pid, uint8_t subtype);
+void xinput_controller_connected(uint16_t vid, uint16_t pid, uint8_t subtype);
 void xone_controller_connected(uint8_t dev_addr);
 void ps4_controller_connected(uint8_t dev_addr, uint16_t vid, uint16_t pid);
 void ps3_controller_connected(uint8_t dev_addr, uint16_t vid, uint16_t pid);
